Create Naive Pattern.cpp
The number of comparisons in the worst case is O(M * (N – M + 1)). Although strings which have repeated characters are not likely to appear in English text, they may well occur in other applications (for example, in binary texts). The KMP matching algorithm improves the worst case to O(N). We will be covering KMP in the next post. Also, we will be writing more posts to cover all pattern searching algorithms and data structures